article edit. Only owner can edit and delete articles. Admin can edit and delete all user's articles

+++ src/app/article/article-list/article-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {Article} from '../article'
 import { ArticleService } from '../article.service';
 import { Router } from '@angular/router';
+import { TokenStorageService } from '../../security/token-storage.service';
 
 @Component({
   selector: 'app-article-list',
@@ -13,11 +14,14 @@ export class ArticleListComponent implements OnInit {
   articles: Article[] = [];
 
   constructor(private articleService: ArticleService,
-    private router: Router) { }
+    private router: Router,
+    private token: TokenStorageService) { }
 
   ngOnInit(): void {
     this.getArticles();
   }
+
+
 
   private getArticles(){
     this.articleService.getArticleList().subscribe(data => {
@@ -25,19 +29,27 @@ export class ArticleListComponent implements OnInit {
     });
   }
 
-  updateArticle(id: number){
-    this.router.navigate(['update-article', id]);
+  updateArticle(id: number, userId: number){
+    if(userId == this.token.getUser().id || this.token.getUser().roles.includes("ROLE_ADMIN") == true ){
+      this.router.navigate(['update-article', id]);
+    }
+    console.log("article did not create by user");
   }
 
-  deleteArticle(id: number){
-    this.articleService.deleteArticle(id).subscribe(data => {
-      console.log(data);
-      this.getArticles();
-    });
+  deleteArticle(id: number, userId:number){
+    if(userId == this.token.getUser().id || this.token.getUser().roles.includes("ROLE_ADMIN") == true ){
+      this.articleService.deleteArticle(id).subscribe(data => {
+        console.log(data);
+        this.getArticles();
+      });
+    }
+    console.log("article did not create by user");
   }
 
   articleDetails(id: number){
     this.router.navigate(['article-details', id])
   }
+
+
 
 }
